Fixes formating and more
