+++ exclude/formating/blocks.ts
@@ -1,0 +1,20 @@
+// File Header:
+
+/*********************************************************************\
+|*|‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾|*|
+|*|	File Summary:													|*|
+|*|		A short summary of the current file.						|*|
+|*|	Sections:														|*|
+|*|	    - Section Title:    Very long discription that takes        |*|
+|*|                         multiple lines for whatever reason.     |*|
+|*|     - Section Title:    Discription.                            |*|
+|*|_________________________________________________________________|*|
+\*********************************************************************/
+
+// Section Header:
+
+/*********************************************************************\
+|*|‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾|*|
+|*|	Section Title:  Discription.	                    			|*|
+|*|_________________________________________________________________|*|
+\*********************************************************************/
