+++ src/lib/System/EventManager/eventConnection.ts
@@ -1,0 +1,22 @@
+import { IEventConnection, IEvent } from "./types";
+
+/**The connection for a spicific callback on an event. */
+export class EventConnection implements IEventConnection {
+	private readonly Event: IEvent;
+	/**@hidden */
+	readonly Callback: Callback;
+	/**@hidden */
+	readonly once: boolean;
+
+	/**Disconnects the callback from its linked event. */
+	Disconnect(): boolean {
+		return this.Event.Disconnect(this);
+	}
+
+	/**@hideconstructor */
+	constructor(callback: Callback, event: IEvent, once: boolean) {
+		this.Callback = callback;
+		this.Event = event;
+		this.once = once;
+	}
+}
