+++ src/lib/System/NetLink/index.ts
@@ -1,0 +1,1 @@
+export namespace Net {}
