+++ src/lib/System/TechTree/world-object/types.d.ts
