--- src/lib/systems/input/types.d.ts
@@ -1,4 +1,0 @@
-export interface ButtonProps {
-	useLeftMouse: boolean;
-	useRightMouse: boolean;
-}
